backend/requirements.txt
Django>=5.2.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
python-dotenv>=1.0.1
Pillow>=10.2.0
pydantic>=2.10.0
drf-spectacular>=0.28.0
ruff>=0.9.0
psycopg2-binary>=2.9.9
email-validator>=2.1.0
pyotp>=2.9.0
requests>=2.32.5
django-filter>=24.2
drf-nested-routers>=0.93.5

django-cloudinary-storage>=0.3.0
cloudinary>=1.36.0

# Real-time (WebSocket + SSE)
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.1.0
django-eventstream>=5.3.0
redis>=5.0.0
pymongo>=4.6.1
dnspython>=2.6.0
google-generativeai>=0.8.0
